Create `ImageCoreGraphicsTests` and move `resize` into it
